restructured conf (part of gen front end modules set), restructured routes by apps, fixed server.js, express.js .  TO DO tests broken
