+++ app/admin-routes/users.server.routes.js
@@ -7,7 +7,7 @@ var passport = require('passport');
 
 module.exports = function(app) {
   // User Routes
-  var users = require('../../app/controllers/users.server.controller');
+  var users = require('../controllers/users.server.controller');
 
   // Setting up the users profile api
   app.route('/users/me').get(users.me);
