+++ app/search-api-routes/opt-items.server.routes.js
@@ -1,8 +1,8 @@
 'use strict';
 
 module.exports = function(app) {
-	var users = require('../../app/controllers/users.server.controller');
-	var optItems = require('../../app/controllers/opt-items.server.controller');
+	var users = require('../controllers/users.server.controller.js');
+	var optItems = require('../controllers/opt-items.server.controller.js');
 
 	// Opt items Routes
 	app.route('/opt-items')
