+++ config/config.js
@@ -74,3 +74,8 @@ module.exports.getCSSAssets = function() {
 	var output = this.getGlobbedFiles(this.assets.lib.css.concat(this.assets.css), 'public/');
 	return output;
 };
+
+/**
+ * additional config
+ */
+module.exports.admin = require('./admin.server.config')(module.exports);
